refactor: v2 Section으로 마이그레이션
