+++ src/main/java/wooteco/subway/domain/Line.java
@@ -7,7 +7,7 @@ public class Line {
     private Long id;
     private String name;
     private String color;
-    private SectionsV2 sections;
+    private Sections sections;
 
     public Line(String name, String color) {
         this(null, name, color);
@@ -19,13 +19,13 @@ public class Line {
         this.color = color;
     }
 
-    public Line(String name, String color, SectionsV2 sections) {
+    public Line(String name, String color, Sections sections) {
         this.id = null;
         this.color = color;
         this.sections = sections;
     }
 
-    public Line(Long id, String name, String color, SectionsV2 sections) {
+    public Line(Long id, String name, String color, Sections sections) {
         this.id = id;
         this.name = name;
         this.color = color;
@@ -44,7 +44,7 @@ public class Line {
         return color;
     }
 
-    public SectionsV2 getSections() {
+    public Sections getSections() {
         return sections;
     }
