+++ src/main/java/wooteco/subway/service/LineService.java
@@ -7,11 +7,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.dao.LineDao;
 import wooteco.subway.dao.SectionDao;
-import wooteco.subway.dao.SectionDaoV2;
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Line;
-import wooteco.subway.domain.SectionV2;
-import wooteco.subway.domain.SectionsV2;
+import wooteco.subway.domain.Section;
+import wooteco.subway.domain.Sections;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.LineRequest;
 import wooteco.subway.dto.LineResponse;
@@ -25,14 +24,11 @@ public class LineService {
     private final StationDao stationDao;
     private final LineDao lineDao;
     private final SectionDao sectionDao;
-    private final SectionDaoV2 sectionDaoV2;
 
-    public LineService(StationDao stationDao, LineDao lineDao, SectionDao sectionDao,
-                       SectionDaoV2 sectionDaoV2) {
+    public LineService(StationDao stationDao, LineDao lineDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
         this.lineDao = lineDao;
         this.sectionDao = sectionDao;
-        this.sectionDaoV2 = sectionDaoV2;
     }
 
     @Transactional
@@ -45,8 +41,8 @@ public class LineService {
         final Station upStation = stationDao.findById(request.getUpStationId());
         final Station downStation = stationDao.findById(request.getDownStationId());
 
-        final SectionV2 section = new SectionV2(lineId, upStation, downStation, request.getDistance());
-        sectionDaoV2.save(section);
+        final Section section = new Section(lineId, upStation, downStation, request.getDistance());
+        sectionDao.save(section);
 
         return lineId;
     }
@@ -77,14 +73,14 @@ public class LineService {
 
     public LineResponse findById(Long id) {
         Line line = lineDao.findById(id);
-        SectionsV2 sections = new SectionsV2(sectionDaoV2.findByLineId(id));
+        Sections sections = new Sections(sectionDao.findByLineId(id));
         return new LineResponse(line.getId(),
                 line.getName(),
                 line.getColor(),
                 sortSections(sections));
     }
 
-    private List<StationResponse> sortSections(SectionsV2 sections) {
+    private List<StationResponse> sortSections(Sections sections) {
         List<StationResponse> stationResponses = new ArrayList<>();
         Station firstStation = sections.findFirstStation();
         stationResponses.add(new StationResponse(firstStation.getId(), firstStation.getName()));
