+++ src/test/java/wooteco/subway/dao/LineDaoTest.java
@@ -3,8 +3,6 @@ package wooteco.subway.dao;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static wooteco.subway.Fixtures.강남역;
-import static wooteco.subway.Fixtures.역삼역;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
@@ -13,13 +11,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import wooteco.subway.domain.Line;
-import wooteco.subway.domain.SectionV2;
+import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
 
 @Import({
         LineDao.class,
         StationDao.class,
-        SectionDaoV2.class
+        SectionDao.class
 })
 @JdbcTest
 class LineDaoTest {
@@ -29,7 +27,7 @@ class LineDaoTest {
     @Autowired
     private StationDao stationDao;
     @Autowired
-    private SectionDaoV2 sectionDao;
+    private SectionDao sectionDao;
 
     @Test
     @DisplayName("노선을 등록할 수 있다.")
@@ -50,8 +48,8 @@ class LineDaoTest {
     @DisplayName("전체 노선을 조회할 수 있다.")
     void findAll() {
         // given
-        Long 강남역_id = stationDao.save(강남역);
-        Long 역삼역_id = stationDao.save(역삼역);
+        Long 강남역_id = stationDao.save(new Station("강남역"));
+        Long 역삼역_id = stationDao.save(new Station("역삼역"));
         Station 강남역 = stationDao.findById(강남역_id);
         Station 역삼역 = stationDao.findById(역삼역_id);
 
@@ -60,8 +58,8 @@ class LineDaoTest {
         Long 신분당선_id = lineDao.save(신분당선);
         Long 분당선_id = lineDao.save(분당선);
 
-        SectionV2 강남_역삼_신분당선 = new SectionV2(신분당선_id, 강남역, 역삼역, 10);
-        SectionV2 강남_역삼_분당선 = new SectionV2(분당선_id, 강남역, 역삼역, 10);
+        Section 강남_역삼_신분당선 = new Section(신분당선_id, 강남역, 역삼역, 10);
+        Section 강남_역삼_분당선 = new Section(분당선_id, 강남역, 역삼역, 10);
         sectionDao.save(강남_역삼_신분당선);
         sectionDao.save(강남_역삼_분당선);
 
@@ -80,14 +78,14 @@ class LineDaoTest {
     @DisplayName("단건 노선을 조회한다.")
     void findById() {
         // given
-        Long 강남역_id = stationDao.save(강남역);
-        Long 역삼역_id = stationDao.save(역삼역);
+        Long 강남역_id = stationDao.save(new Station("강남역"));
+        Long 역삼역_id = stationDao.save(new Station("역삼역"));
         Station 강남역 = stationDao.findById(강남역_id);
         Station 역삼역 = stationDao.findById(역삼역_id);
 
         Line 신분당선 = new Line("신분당선", "bg-red-600");
         Long 신분당선_id = lineDao.save(신분당선);
-        sectionDao.save(new SectionV2(신분당선_id, 강남역, 역삼역, 10));
+        sectionDao.save(new Section(신분당선_id, 강남역, 역삼역, 10));
 
         // when
         Line findLine = lineDao.findById(신분당선_id);
